Authentication feature: Moved to Gin + created User and Create Table functions  (#16)
Overview of Changes:
This PR is part 1 of 2 of creating a authentication feature, to create a authentication feature, we need a users table and employees table created for teams. In this iteration I created the specified tables, switched to Gin for easier backend decoupling from GET and POST requests. The next iteration will tackle User creation and JWT token auth
---------

Co-authored-by: Michael Lee <mlee@MacBook-Pro.local>